test(release): split oversized workflow coverage
